micro_service_ocr_tickets_GastApp/requirements.txt
fastapi
uvicorn
transformers
torch
accelerate
Pillow
pytesseract
python-multipart
